// app/src/generated/models/createPaymentInstructionRequestPaymentMethod.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Supported payment method for rent collection
 */
export type CreatePaymentInstructionRequestPaymentMethod =
  (typeof CreatePaymentInstructionRequestPaymentMethod)[keyof typeof CreatePaymentInstructionRequestPaymentMethod];

export const CreatePaymentInstructionRequestPaymentMethod = {
  BANK_TRANSFER: 'BANK_TRANSFER',
  PAYPAL: 'PAYPAL',
  CASH: 'CASH',
  CHECK: 'CHECK',
  DIRECT_DEBIT: 'DIRECT_DEBIT',
  IDEAL_WERO: 'IDEAL_WERO',
  ZELLE: 'ZELLE',
  OTHER: 'OTHER',
} as const;